requirements.txt
streamlit
google-generativeai
transformers
sentencepiece
torch
gtts
speechrecognition
python-dotenv
langchain
langchain-community
pypdf
faiss-cpu
sentence-transformers
huggingface-hub
